pyproject.toml
[project]
name = "coinmarket"
version = "0.1.0"
description = ""
authors = [
    {name = "phenix3443",email = "phenix3443@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ccxt (>=4.4.77,<5.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "numpy (>=2.2.5,<3.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "yagmail (>=0.15.293,<0.16.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "flask (>=3.1.0,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
